#70 Statusをリファクタリング
